feat: Implement method to retrieve user's joined teams in Microsoft Teams integration
